place config in a common package
